Ianhelle/tests and tidy 2021 09 01 (#200)
* Added separate unit test for nbwidgets to cover cases missed in notebook.

Couple of minor fixes for nbwidgets.py

* Removing deprecated code from utility.py

Removing unneeded function from mordor_driver.py
Sourcery code refactor in entity.py
Deprecating create_ip_record function in ip_utils.py
Split wsconfig tests out of test_pkg_config.py
Adding test_wsconfig.py. Added new test for couple of utility methods
Removing redundant import from test_nbwidgets.py

* Moved lots of test files into sub-package folders corresponding to tested modules

Fixing up paths and removing unneeded imports and other cruft.
Merged a few tests into single files - test_entity.py and test_timeline.py

* Re-enabling IP_stack/geoip notebook test in test_geoip.py

Updating use of tqdm to remove deprecated tqdm_notebook in utility.py

* Adding some debugging output to geoip GeoLite class.

Suppressing mypy warning in azure_ml_tools.py

* Updating test_geoip.py to use predictable msticpyconfig-test.yaml

* Adding debug statement to show config file and partial API key

* Typo in last push to geoip.py

* Another typo in the debug code in geoip.py

* Updating test to restore env var values after test in test_pkg_config.py

* More debugging for test_geoip.py

Fixing aliases element in kql query records
Handle aliases as list or string in help

* Fixing some merge errors and test failures in geoip.py, import_analyzer.py, test_azuredata.py, test_geoip.py,test_azuredatatest_process_tree_utils.py, test_user_config.py

+++ tools/toollib/import_analyzer.py
@@ -9,6 +9,7 @@ import re
 import sys
 from importlib import import_module
 from typing import Dict, List, Optional, Set
+import warnings
 
 import networkx as nx
 
@@ -135,8 +136,7 @@ def get_extras_from_setup(
 
     neut_setup_py = Path(package_root) / "neut_setup.py"
     try:
-        with open(neut_setup_py, "w") as f_handle:
-            f_handle.writelines(setup_txt)
+        neut_setup_py.write_text(setup_txt)
 
         setup_mod = import_module("neut_setup")
         extras = getattr(setup_mod, "EXTRAS").get(extra)
@@ -147,7 +147,14 @@ def get_extras_from_setup(
             )
         return sorted(list(set(extras)), key=str.casefold)
     finally:
-        neut_setup_py.unlink()
+        try:
+            if neut_setup_py.is_file():
+                try:
+                    neut_setup_py.unlink()
+                except (FileNotFoundError, PermissionError) as err:
+                    warnings.warn(f"Unable to remove {neut_setup_py}. Error {err}")
+        except PermissionError:
+            print(f"could not remove temp file {neut_setup_py}")
 
 
 def _extract_pkg_specs(pkg_specs: List[str]):
